# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Socle SIRH — Sorbonne-Hôtel",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/besoins": {
      "get": {
        "tags": [
          "Besoins identifiés"
        ],
        "summary": "Les besoins identifiés, filtrables par type et par statut",
        "operationId": "lister_besoins_besoins_get",
        "security": [
          {
            "Jeton de groupe": []
          }
        ],
        "parameters": [
          {
            "name": "type_besoin",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/NeedType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "recrutement, formation ou mobilite",
              "title": "Type Besoin"
            },
            "description": "recrutement, formation ou mobilite"
          },
          {
            "name": "statut",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/NeedStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Statut"
            }
          },
          {
            "name": "priorite",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/NeedPriority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priorite"
            }
          },
          {
            "name": "etablissement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Etablissement Id"
            }
          },
          {
            "name": "departement_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Departement Id"
            }
          },
          {
            "name": "poste_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Poste Id"
            }
          },
          {
            "name": "salarie_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Salarie Id"
            }
          },
          {
            "name": "competence_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Competence Id"
            }
          },
          {
            "name": "module",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Le code du groupe ayant pris le besoin en charge",
              "title": "Module"
            },
            "description": "Le code du groupe ayant pris le besoin en charge"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BesoinOut"
                  },
                  "title": "Response Lister Besoins Besoins Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/besoins/{besoin_id}": {
      "get": {
        "tags": [
          "Besoins identifiés"
        ],
        "summary": "Un besoin identifié",
        "operationId": "lire_besoin_besoins__besoin_id__get",
        "security": [
          {
            "Jeton de groupe": []
          }
        ],
        "parameters": [
          {
            "name": "besoin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Besoin Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BesoinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Besoins identifiés"
        ],
        "summary": "Faire évoluer le statut d'un besoin — la seule écriture autorisée sur le socle",
        "description": "Réservé aux trois modules qui consomment des besoins, et chacun ne traite\nque son type : Recrutement les besoins de recrutement, Formation ceux de\nformation, Mobilité ceux de mobilité.\n\nLe module n'est pas à renseigner : il est déduit du jeton.",
        "operationId": "modifier_besoin_besoins__besoin_id__patch",
        "security": [
          {
            "Jeton de groupe": []
          }
        ],
        "parameters": [
          {
            "name": "besoin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Besoin Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BesoinPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BesoinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BesoinOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "need_type": {
            "$ref": "#/components/schemas/NeedType"
          },
          "need_type_label": {
            "type": "string",
            "title": "Need Type Label"
          },
          "priority": {
            "$ref": "#/components/schemas/NeedPriority"
          },
          "priority_label": {
            "type": "string",
            "title": "Priority Label"
          },
          "status": {
            "$ref": "#/components/schemas/NeedStatus"
          },
          "status_label": {
            "type": "string",
            "title": "Status Label"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "detected_on": {
            "type": "string",
            "format": "date",
            "title": "Detected On"
          },
          "employee_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Id"
          },
          "employee_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Name"
          },
          "matricule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matricule"
          },
          "position_id": {
            "type": "integer",
            "title": "Position Id"
          },
          "position_title": {
            "type": "string",
            "title": "Position Title"
          },
          "job_family": {
            "type": "string",
            "title": "Job Family"
          },
          "hierarchy_level": {
            "type": "integer",
            "title": "Hierarchy Level"
          },
          "department_id": {
            "type": "integer",
            "title": "Department Id"
          },
          "department_label": {
            "type": "string",
            "title": "Department Label"
          },
          "establishment_id": {
            "type": "integer",
            "title": "Establishment Id"
          },
          "establishment_name": {
            "type": "string",
            "title": "Establishment Name"
          },
          "establishment_city": {
            "type": "string",
            "title": "Establishment City"
          },
          "skill_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Id"
          },
          "skill_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Code"
          },
          "skill_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Label"
          },
          "review_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Id"
          },
          "handled_by_module": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handled By Module"
          },
          "handled_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handled On"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "need_type",
          "need_type_label",
          "priority",
          "priority_label",
          "status",
          "status_label",
          "origin",
          "detected_on",
          "employee_id",
          "employee_name",
          "matricule",
          "position_id",
          "position_title",
          "job_family",
          "hierarchy_level",
          "department_id",
          "department_label",
          "establishment_id",
          "establishment_name",
          "establishment_city",
          "skill_id",
          "skill_code",
          "skill_label",
          "review_id",
          "handled_by_module",
          "handled_on",
          "comment"
        ],
        "title": "BesoinOut"
      },
      "BesoinPatch": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/NeedStatus"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "BesoinPatch",
        "description": "La seule écriture autorisée aux groupes sur le socle.\n\nLe module qui prend le besoin en charge n'est pas à renseigner : il est\ndéduit du jeton, pour qu'un groupe ne puisse pas écrire au nom d'un autre."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NeedPriority": {
        "type": "string",
        "enum": [
          "haute",
          "moyenne",
          "basse"
        ],
        "title": "NeedPriority"
      },
      "NeedStatus": {
        "type": "string",
        "enum": [
          "ouvert",
          "pris_en_charge",
          "cloture"
        ],
        "title": "NeedStatus"
      },
      "NeedType": {
        "type": "string",
        "enum": [
          "recrutement",
          "formation",
          "mobilite"
        ],
        "title": "NeedType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "Jeton de groupe": {
        "type": "http",
        "description": "Le jeton remis à votre groupe, en en-tête `Authorization: Bearer <jeton>`.",
        "scheme": "bearer"
      }
    }
  }
}